+++ src/main/java/com/server/objet/global/client/kakao/service/dto/response/KakaoAccount.java
@@ -1,4 +1,4 @@
-package com.server.objet.global.client.kakao.dto.response;
+package com.server.objet.global.client.kakao.service.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,8 +9,7 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoUserProfile {
+public class KakaoAccount {
 
-    private String nickname;
-    private String profileImageUrl;
+    private KakaoUserProfile profile;
 }
